src/familiar_agent/config.py
"""Configuration management."""

from __future__ import annotations

import os
from dataclasses import dataclass, field
from pathlib import Path

from dotenv import load_dotenv

load_dotenv()


@dataclass
class CameraConfig:
    host: str = field(
        default_factory=lambda: os.environ.get(
            "CAMERA_HOST", os.environ.get("TAPO_CAMERA_HOST", "")
        )
    )
    username: str = field(
        default_factory=lambda: os.environ.get(
            "CAMERA_USERNAME", os.environ.get("TAPO_USERNAME", "admin")
        )
    )
    password: str = field(
        default_factory=lambda: os.environ.get(
            "CAMERA_PASSWORD", os.environ.get("TAPO_PASSWORD", "")
        )
    )
    port: int = field(
        default_factory=lambda: int(
            os.environ.get("CAMERA_ONVIF_PORT", os.environ.get("TAPO_ONVIF_PORT", "2020"))
        )
    )


@dataclass
class MobilityConfig:
    api_region: str = field(default_factory=lambda: os.environ.get("TUYA_REGION", "us"))
    api_key: str = field(default_factory=lambda: os.environ.get("TUYA_API_KEY", ""))
    api_secret: str = field(default_factory=lambda: os.environ.get("TUYA_API_SECRET", ""))
    device_id: str = field(default_factory=lambda: os.environ.get("TUYA_DEVICE_ID", ""))


@dataclass
class TTSConfig:
    elevenlabs_api_key: str = field(
        default_factory=lambda: os.environ.get("ELEVENLABS_API_KEY", "")
    )
    voice_id: str = field(
        default_factory=lambda: os.environ.get("ELEVENLABS_VOICE_ID", "cgSgspJ2msm6clMCkdW9")
    )
    go2rtc_url: str = field(
        default_factory=lambda: os.environ.get("GO2RTC_URL", "http://localhost:1984")
    )
    go2rtc_stream: str = field(default_factory=lambda: os.environ.get("GO2RTC_STREAM", "tapo_cam"))


@dataclass
class MemoryConfig:
    db_path: str = field(
        default_factory=lambda: os.environ.get(
            "MEMORY_DB_PATH",
            str(Path.home() / ".claude" / "memories"),
        )
    )


@dataclass
class AgentConfig:
    # Agent display name shown in TUI
    agent_name: str = field(default_factory=lambda: os.environ.get("AGENT_NAME", "AI"))

    # Platform: "gemini" | "anthropic" | "openai"
    platform: str = field(default_factory=lambda: os.environ.get("PLATFORM", "anthropic"))

    # Unified API key (used for whichever platform is selected)
    api_key: str = field(default_factory=lambda: os.environ.get("API_KEY", ""))

    # Model name — platform-specific defaults applied in create_backend()
    model: str = field(default_factory=lambda: os.environ.get("MODEL", ""))

    # OpenAI-compatible only: base URL and tool-calling mode
    # TOOLS_MODE: "native" = use function-calling API, "prompt" = inject into system prompt
    base_url: str = field(
        default_factory=lambda: os.environ.get("BASE_URL", "http://localhost:11434/v1")
    )
    tools_mode: str = field(default_factory=lambda: os.environ.get("TOOLS_MODE", "prompt"))

    max_tokens: int = 4096
    camera: CameraConfig = field(default_factory=CameraConfig)
    mobility: MobilityConfig = field(default_factory=MobilityConfig)
    tts: TTSConfig = field(default_factory=TTSConfig)
    memory: MemoryConfig = field(default_factory=MemoryConfig)
